gm: refactor upstream controller
Split upstream controller, merge each feature CR logic code into its
controller.
