+++ pkg/globalmanager/controllers/federatedlearning/federatedlearningjob.go
@@ -18,6 +18,7 @@ package federatedlearning
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"strconv"
 	"time"
@@ -51,6 +52,9 @@ import (
 )
 
 const (
+	// KindName is the kind name of CR this controller controls
+	KindName = "FederatedLearningJob"
+	// Name is this controller name
 	Name = "FederatedLearning"
 )
 
@@ -60,7 +64,7 @@ const (
 )
 
 // Kind contains the schema.GroupVersionKind for this controller type.
-var Kind = sednav1.SchemeGroupVersion.WithKind("FederatedLearningJob")
+var Kind = sednav1.SchemeGroupVersion.WithKind(KindName)
 
 // Controller ensures that all FLJob objects have corresponding pods to
 // run their configured workload.
@@ -531,9 +535,102 @@ func (c *Controller) createPod(job *sednav1.FederatedLearningJob) (active int32,
 	return
 }
 
+func (c *Controller) updateModelMetrics(jobName, namespace string, metrics []sednav1.Metric) error {
+	var err error
+	job, err := c.client.FederatedLearningJobs(namespace).Get(context.TODO(), jobName, metav1.GetOptions{})
+	if err != nil {
+		// federated crd not found
+		return err
+	}
+	modelName := job.Spec.AggregationWorker.Model.Name
+	client := c.client.Models(namespace)
+
+	return runtime.RetryUpdateStatus(modelName, namespace, (func() error {
+		model, err := client.Get(context.TODO(), modelName, metav1.GetOptions{})
+		if err != nil {
+			return err
+		}
+
+		now := metav1.Now()
+		model.Status.UpdateTime = &now
+		model.Status.Metrics = metrics
+		_, err = client.UpdateStatus(context.TODO(), model, metav1.UpdateOptions{})
+		return err
+	}))
+}
+
+func (c *Controller) appendStatusCondition(name, namespace string, cond sednav1.FLJobCondition) error {
+	client := c.client.FederatedLearningJobs(namespace)
+
+	return runtime.RetryUpdateStatus(name, namespace, (func() error {
+		job, err := client.Get(context.TODO(), name, metav1.GetOptions{})
+		if err != nil {
+			return err
+		}
+		job.Status.Conditions = append(job.Status.Conditions, cond)
+		_, err = client.UpdateStatus(context.TODO(), job, metav1.UpdateOptions{})
+		return err
+	}))
+}
+
+// updateFromEdge updates the federated job's status
+func (c *Controller) updateFromEdge(name, namespace, operation string, content []byte) (err error) {
+	// JobInfo defines the job information
+	type JobInfo struct {
+		// Current training round
+		CurrentRound int    `json:"currentRound"`
+		UpdateTime   string `json:"updateTime"`
+	}
+
+	// Output defines job output information
+	type Output struct {
+		Models  []runtime.Model `json:"models"`
+		JobInfo *JobInfo        `json:"ownerInfo"`
+	}
+
+	var status struct {
+		Phase  string  `json:"phase"`
+		Status string  `json:"status"`
+		Output *Output `json:"output"`
+	}
+
+	err = json.Unmarshal(content, &status)
+	if err != nil {
+		return
+	}
+
+	output := status.Output
+
+	if output != nil {
+		// Update the model's metrics
+		if len(output.Models) > 0 {
+			// only one model
+			model := output.Models[0]
+			metrics := runtime.ConvertMapToMetrics(model.Metrics)
+			if len(metrics) > 0 {
+				c.updateModelMetrics(name, namespace, metrics)
+			}
+		}
+
+		jobInfo := output.JobInfo
+		// update job info if having any info
+		if jobInfo != nil && jobInfo.CurrentRound > 0 {
+			// Find a good place to save the progress info
+			// TODO: more meaningful reason/message
+			reason := "DoTraining"
+			message := fmt.Sprintf("Round %v reaches at %s", jobInfo.CurrentRound, jobInfo.UpdateTime)
+			cond := NewFLJobCondition(sednav1.FLJobCondTraining, reason, message)
+			c.appendStatusCondition(name, namespace, cond)
+		}
+	}
+
+	return nil
+}
+
 // New creates a new federated learning job controller that keeps the relevant pods
 // in sync with their corresponding FederatedLearningJob objects.
-func New(cfg *config.ControllerConfig) (runtime.FeatureControllerI, error) {
+func New(controllerContext *runtime.ControllerContext) (runtime.FeatureControllerI, error) {
+	cfg := controllerContext.Config
 	namespace := cfg.Namespace
 	if namespace == "" {
 		namespace = metav1.NamespaceAll
@@ -585,5 +682,8 @@ func New(cfg *config.ControllerConfig) (runtime.FeatureControllerI, error) {
 	stopCh := make(chan struct{})
 	kubeInformerFactory.Start(stopCh)
 	jobInformerFactory.Start(stopCh)
+
+	controllerContext.UpstreamController.Add(KindName, fc.updateFromEdge)
+
 	return fc, err
 }
