+++ pkg/globalmanager/controllers/jointinference/jointinferenceservice.go
@@ -52,7 +52,11 @@ import (
 )
 
 const (
+	// Name is this controller name
 	Name = "JointInference"
+
+	// KindName is the kind name of CR this controller controls
+	KindName = "JointInferenceService"
 )
 
 const (
@@ -62,7 +66,7 @@ const (
 )
 
 // Kind contains the schema.GroupVersionKind for this controller type.
-var Kind = sednav1.SchemeGroupVersion.WithKind("JointInferenceService")
+var Kind = sednav1.SchemeGroupVersion.WithKind(Name)
 
 // Controller ensures that all JointInferenceService objects
 // have corresponding pods to run their configured workload.
@@ -537,9 +541,70 @@ func (c *Controller) createEdgeWorker(service *sednav1.JointInferenceService, bi
 	return err
 }
 
+func (c *Controller) updateMetrics(name, namespace string, metrics []sednav1.Metric) error {
+	client := c.client.JointInferenceServices(namespace)
+
+	return runtime.RetryUpdateStatus(name, namespace, func() error {
+		joint, err := client.Get(context.TODO(), name, metav1.GetOptions{})
+		if err != nil {
+			return err
+		}
+		joint.Status.Metrics = metrics
+		_, err = client.UpdateStatus(context.TODO(), joint, metav1.UpdateOptions{})
+		return err
+	})
+}
+
+// updateFromEdge syncs the edge updates to k8s
+func (c *Controller) updateFromEdge(name, namespace, operation string, content []byte) error {
+	// Output defines owner output information
+	type Output struct {
+		ServiceInfo map[string]interface{} `json:"ownerInfo"`
+	}
+
+	var status struct {
+		// Phase always should be "inference"
+		Phase  string  `json:"phase"`
+		Status string  `json:"status"`
+		Output *Output `json:"output"`
+	}
+
+	err := json.Unmarshal(content, &status)
+	if err != nil {
+		return err
+	}
+
+	// TODO: propagate status.Status to k8s
+
+	output := status.Output
+	if output == nil || output.ServiceInfo == nil {
+		// no output info
+		klog.Warningf("empty status info for joint inference service %s/%s", namespace, name)
+		return nil
+	}
+
+	info := output.ServiceInfo
+
+	for _, ignoreTimeKey := range []string{
+		"startTime",
+		"updateTime",
+	} {
+		delete(info, ignoreTimeKey)
+	}
+
+	metrics := runtime.ConvertMapToMetrics(info)
+
+	err = c.updateMetrics(name, namespace, metrics)
+	if err != nil {
+		return fmt.Errorf("failed to update metrics, err:%+w", err)
+	}
+	return nil
+}
+
 // New creates a new JointInferenceService controller that keeps the relevant pods
 // in sync with their corresponding JointInferenceService objects.
-func New(cfg *config.ControllerConfig) (runtime.FeatureControllerI, error) {
+func New(controllerContext *runtime.ControllerContext) (runtime.FeatureControllerI, error) {
+	cfg := controllerContext.Config
 	var err error
 	namespace := cfg.Namespace
 	if namespace == "" {
@@ -597,5 +662,8 @@ func New(cfg *config.ControllerConfig) (runtime.FeatureControllerI, error) {
 	stopCh := messageContext.Done()
 	kubeInformerFactory.Start(stopCh)
 	serviceInformerFactory.Start(stopCh)
+
+	controllerContext.UpstreamController.Add(KindName, jc.updateFromEdge)
+
 	return jc, err
 }
